requirements.txt
langchain
python-dotenv==1.0.0
streamlit
streamlit_chat==0.0.2.2
langchain_community
